+++ routes/beneficiary.routes.js
@@ -1,0 +1,9 @@
+const express = require("express");
+const router = express.Router();
+
+const beneficiary_controller = require('../controllers/beneficiary.controller');
+
+router.get('/', beneficiary_controller.get_beneficiaries);
+router.get('/:id', beneficiary_controller.get_beneficiary);
+
+module.exports = router;