+++ CHANGELOG.md
@@ -1,7 +1,35 @@
+## [0.0.0] - 01/11/2025  
+
+### Changed
+ - Modified beneficiary model to correctly fetch data from the database using async/await.
+ - Updated beneficiary controller to handle errors when fetching beneficiary by ID.
+
+## [0.0.0] - 31/10/2025  
+
+### Changed
+
+ - Updated dotenv configuration to correctly load environment variables from the .env file located in the parent directory.
+
+### Added
+ - Test code to verify database connection and query execution.
+ - Added "type": "module" to package.json to enable ES module support.
+
+## [0.0.0] - 20/10/2025  
+
+### Added
+
+ - MC (Model-Controller) and routes (endpoints)
+ - Mock data to verify the correct implementation of the API
+ - APIs initial structure with NodeJs and Express
+
 ## [0.0.0] - 14/10/2025
 
 ### Added
 
+ - Coding Standard (14/10/2025)
  - Project's initial commit (14/10/2025)
  - Changelog file (14/10/2025)
  - Pull request template file (14/10/2025)
+
+### Changed
+ - Pull request template file now has a direct link to the coding standard (21/10/2025)